carpeta para los themes personalizados de Chakra
