+++ frontend/src/App.tsx
@@ -11,13 +11,14 @@ import AdmPage from "./pages/AdmPage/AdmPage";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import EstablecimientosPage from "./pages/EstablecimientosPage/EstablecimientosPage";
 import { CurrentAdminProvider } from "./hooks/useCurrentAdmin";
-import { ChakraProvider, extendTheme, Theme } from "@chakra-ui/react";
+import { ChakraProvider } from "@chakra-ui/react";
 import EditEstab from "./pages/EditEstab/EditEstab";
 import PerfilPage from "./pages/PerfilPage/PerfilPage";
 import NewCourt from "./pages/NewCourt/NewCourt";
 import CourtPage from "./pages/CourtPage/CourtPage";
 import NotFound from "./pages/NotFoundPage/NotFound";
 import EditCourt from "./pages/EditCourt/EditCourt";
+import { theme } from "./themes";
 
 const routes = [
   {
@@ -73,54 +74,6 @@ const routes = [
     element: <EditCourt />,
   },
 ];
-
-export const theme: Partial<Theme> = extendTheme({
-  components: {
-    Input: {
-      variants: {
-        floating: {
-          field: {
-            "&::placeholder": {
-              color: "gray",
-            },
-          },
-        },
-      },
-    },
-    Form: {
-      variants: {
-        floating: {
-          container: {
-            _focusWithin: {
-              label: {
-                transform: "scale(0.85) translateY(-24px)",
-                fontWeight: "normal",
-              },
-            },
-            "input:not(:placeholder-shown) + label, .chakra-select__wrapper + label, textarea:not(:placeholder-shown) ~ label":
-              {
-                transform: "scale(0.85) translateY(-24px)",
-                fontWeight: "normal",
-              },
-            label: {
-              top: 0,
-              left: 0,
-              zIndex: 2,
-              position: "absolute",
-              backgroundColor: "white",
-              pointerEvents: "none",
-              mx: 3,
-              px: 1,
-              my: 2,
-              transformOrigin: "left top",
-              fontWeight: "normal",
-            },
-          },
-        },
-      },
-    },
-  },
-});
 
 const router = createBrowserRouter(routes);
 const queryClient = new QueryClient();
